payback_mlop_demo_full/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
joblib
pytest
python-dotenv
google-cloud-storage
sqlalchemy
